move logging to rhcephcompose.log
This fixes a flake8 error in __init__.py, "E402 module level import not
at top of file"
