+++ rhcephcompose/__init__.py
@@ -1,8 +1,3 @@
-import logging
-
-logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.INFO)
-log = logging.getLogger('rhcephcompose')
-
 from .build import Build
 from .comps import Comps
 from .variants import Variants
