+++ rhcephcompose/log.py
@@ -1,0 +1,4 @@
+import logging
+
+logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.INFO)
+log = logging.getLogger('rhcephcompose')
